MAGETWO-31949: Error message after successful install

+++ setup/module/Magento/Setup/src/Model/WebLogger.php
@@ -164,7 +164,9 @@ class WebLogger implements LoggerInterface
      */
     public function clear()
     {
-        unlink($this->logFile);
+        if (file_exists($this->logFile)) {
+            unlink($this->logFile);
+        }
     }
 
     /**
